pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <your@email.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.7"
